Crate rattler_package_streaming

This crate provides the ability to extract a Conda package archive or specific parts of it.

Modules

fs

Functions to extracting or stream a Conda package from a file on disk.

read

Functions that enable extracting or streaming a Conda package for objects that implement the std::io::Read trait.

reqwest

Functionality to stream and extract packages directly from a reqwest::Url.

seek

Functionality to stream parts of a .conda archive for objects that implement both std::io::Read and std::io::Seek like a std::fs::File or a std::io::Cursor<T>.

tokio

Functionality to stream and extract packages in an [tokio] async context.

write

Functionality for writing conda packages

Structs

ExtractResult

Result struct returned by extraction functions.

Enums

ExtractError

An error that can occur when extracting a package archive.

Traits

DownloadReporter

A trait that can be implemented to report download progress.