rattler_package_streaming::reqwest::tokio

Function extract

pub async fn extract(
    client: ClientWithMiddleware,
    url: Url,
    destination: &Path,
    expected_sha256: Option<Sha256Hash>,
    reporter: Option<Arc<dyn DownloadReporter>>,
) -> Result<ExtractResult, ExtractError>

Extracts the contents a package archive from the specified remote location. The type of package is determined based on the path of the url.

use url::Url;
use rattler_package_streaming::reqwest::tokio::extract;
use reqwest::Client;
use reqwest_middleware::ClientWithMiddleware;
let _ = extract(
    ClientWithMiddleware::from(Client::new()),
    Url::parse("https://conda.anaconda.org/conda-forge/linux-64/python-3.10.8-h4a9ceb5_0_cpython.conda").unwrap(),
    Path::new("/tmp"),
    None,
    None)
    .await
    .unwrap();