rattler_package_streaming

Module seek

Functionality to stream parts of a .conda archive for objects that implement both std::io::Read and std::io::Seek like a std::fs::File or a std::io::Cursor<T>.

Functions

extract_package_file

Get a PackageFile from temporary archive and extract it to a writer

read_package_file

Read a package file from archive Note: If you want to extract multiple info/* files then this will be slightly slower than manually iterating over the archive entries with custom logic as this skips over the rest of the archive

read_package_file_content

Read a package file content from archive based on the path

stream_conda_content

Stream the content section of a .conda package as a tar archive.

stream_conda_info

Stream the info section of a .conda package as a tar archive.