rattler_package_streaming::seek

Function read_package_file

pub fn read_package_file<P: PackageFile>(
    path: impl AsRef<Path>,
) -> Result<P, ExtractError>

Read a package file from archive Note: If you want to extract multiple info/* files then this will be slightly slower than manually iterating over the archive entries with custom logic as this skips over the rest of the archive

Example

use rattler_conda_types::package::AboutJson;
use rattler_package_streaming::seek::read_package_file;

let about_json = read_package_file::<AboutJson>("conda-forge/win-64/python-3.11.0-hcf16a7b_0_cpython.conda").unwrap();