rattler_package_streaming::tokio

Module async_read

Functions that enable extracting or streaming a Conda package for objects that implement the [tokio::io::AsyncRead] trait.

Functions

extract_conda

Extracts the contents of a .conda package archive using fully async implementation. This will perform on-the-fly decompression by streaming the reader.

extract_conda_via_buffering

Extracts the contents of a .conda package archive by fully reading the stream and then decompressing. This is a fallback method for when streaming fails.

extract_tar_bz2

Extracts the contents a .tar.bz2 package archive using fully async implementation.