rattler_package_streaming

Module write

Functionality for writing conda packages

Traits

ProgressBar

Trait for progress bars

Functions

write_conda_package

Write a .conda package to a writer A .conda package is an outer uncompressed zip archive that contains a metadata.json file, a pkg-archive.tar.zst file, and a info-archive.tar.zst file. The inner zstd encoded files are sorted alphabetically. The info-archive.tar.zst file comes last in the outer zip archive.

write_tar_bz2_package

Write the contents of a list of paths to a tar.bz2 package The paths are sorted alphabetically, and paths beginning with info/ come first.