rattler_package_streaming::write

Function write_tar_bz2_package

pub fn write_tar_bz2_package<W: Write>(
    writer: W,
    base_path: &Path,
    paths: &[PathBuf],
    compression_level: CompressionLevel,
    timestamp: Option<&DateTime<Utc>>,
    progress_bar: Option<Box<dyn ProgressBar>>,
) -> Result<(), Error>

Write the contents of a list of paths to a tar.bz2 package The paths are sorted alphabetically, and paths beginning with info/ come first.

Arguments

writer - the writer to write the package to

base_path - the base path of the package. All paths in paths are relative to this path

paths - a list of paths to include in the package

compression_level - the compression level to use for the inner bzip2 encoded files

timestamp - optional a timestamp to use for all archive files (useful for reproducible builds)

Errors

This function will return an error if the writer returns an error, or if the paths are not relative to the base path.

Examples

use std::path::PathBuf;
use std::fs::File;
use rattler_package_streaming::write::{write_tar_bz2_package};
use rattler_conda_types::compression_level::CompressionLevel;

let paths = vec![PathBuf::from("info/recipe/meta.yaml"), PathBuf::from("info/recipe/conda_build_config.yaml")];
let mut file = File::create("test.tar.bz2").unwrap();
write_tar_bz2_package(&mut file, &PathBuf::from("test"), &paths, CompressionLevel::Default, None, None).unwrap();

See also

write_conda_package